Fix compiler errors on MSVC

R=gauravsh@chromium.org
BUG=none
TEST=make && make runtests

Review URL: http://codereview.chromium.org/6676039

+++ firmware/lib/vboot_firmware.c
@@ -54,7 +54,7 @@ int LoadFirmware(LoadFirmwareParams* params) {
   uint32_t status;
   uint32_t test_err = 0;
   int good_index = -1;
-  int boot_fw_keyblock_flags = 0;
+  uint64_t boot_fw_keyblock_flags = 0;
   int is_dev;
   int index;
   int i;
@@ -292,7 +292,7 @@ int LoadFirmware(LoadFirmwareParams* params) {
   /* At this point, we have a good idea of how we are going to boot. Update the
    * TPM with this state information.
    */
-  status = SetTPMBootModeState(is_dev, 0, boot_fw_keyblock_flags);
+  status = SetTPMBootModeState(is_dev, 0, (int)boot_fw_keyblock_flags);
   if (0 != status) {
     VBDEBUG(("Unable to update the TPM with boot mode information.\n"));
     if (status == TPM_E_MUST_REBOOT)
